libjpeg wrapper: Don't switch hook pointers when using JPU mode
Since multiple context are simultaneously supported, we need per context
mode switching between JPU mode and libjpeg mode.  The existing method
would require a context lookup for every call into the libjpeg library.
Replace it with a flag that is only checked for the functions that implement
the JPU interface.